refactor a bunch
