+++ src/lib/models/mentions/index.ts
@@ -1,0 +1,1 @@
+export * from './mentions'
